+++ CHANGELOG.md
@@ -31,4 +31,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Url schemes error
 - Corrected url for UniversalLinks
 - Fixed a bug with different scopes for different flows
+- Fixed a bug when using login flow with permission